requirements.txt
python-telegram-bot==21.3
selenium==4.21.0
pandas==2.2.2
openpyxl==3.1.4
